move scheduler logic to service
